refactor: move project finding logic to dir
As dir module is responsible for navigating project directory contents
it should be the case that it also contains the logic for looking up
the file tree for 'parent' projects.
